requirements.txt
about-time==4.2.1
    #   alive-progress
alive-progress==3.0.1
attrs==22.2.0
    #   flake8-bugbear
black==23.1.0
build==0.10.0
    #   pip-tools
certifi==2022.12.7
    #   requests
charset-normalizer==2.1.1
    #   requests
click==8.1.3
    #   black
    #   pip-tools
colorama==0.4.6
    #   build
    #   click
flake8==6.0.0
    #   flake8-bugbear
    #   flake8-print
flake8-bugbear==23.1.20
flake8-print==5.0.0
grapheme==0.6.0
    #   alive-progress
idna==3.4
    #   requests
isort==5.11.4
mccabe==0.7.0
    #   flake8
mypy-extensions==0.4.3
    #   black
packaging==23.0
    #   black
    #   build
pathspec==0.11.0
    #   black
pip-tools==6.12.2
platformdirs==2.6.2
    #   black
pycodestyle==2.10.0
    #   flake8
    #   flake8-print
pyflakes==3.0.1
    #   flake8
pyproject-hooks==1.0.0
    #   build
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
schedule==1.1.0
urllib3==1.26.14
    #   requests
wheel==0.38.4